Mserror.LimiterError exception (mscmp_syst_limiter v0.1.0)

Defines errors and related metadata for MscmpSystLimiter component errors.

Types

kinds()

@type kinds() :: :delete_counter | :inspect_counter | :check_counter

Represents the available Kinds of error allowed by this Error type.

Available Kinds

:check_counter - Failure checking and incrementing the rate limit.

:inspect_counter - Failure inspecting the rate limit counter.

:delete_counter - Failure deleting the rate limit counter.

t()

@type t() :: %Mserror.LimiterError{
  __exception__: true,
  __mscomponent__: module(),
  __mserror__: true,
  cause: t() | Exception.t() | term() | nil,
  context: MscmpSystError.Types.Context.t() | nil,
  kind: kinds(),
  message: String.t()
}

Describes the structure of an error of generated by this exception.

The cause attribute allows for errors to be nested.

Attributes

The following attributes are those which are set or modified by the user when working with errors.

kind - The kind of error, see kinds/0 for the available Kinds.

message - A string describing the error.

context - contextual information for better understanding the error. If provided, the context should be of type MscmpSystError.Types.Context.t/0.

cause - The cause of the error, which can be an exception, an error tuple, or any other term.

Internal Attributes

These attributes have their values set at compile time and are used internally by the MscmpSystError framework. They are not to be altered by user code. Reading these attributes in user code is safe, however.

__mserror__ - Always true, indicates that this error is compatible with the MscmpSystError framework.

__mscomponent__ - The component that generated the error.

Functions

new(kind, message, opts \\ [])

@spec new(kind :: kinds(), message :: String.t(), opts :: keyword()) :: t()

Creates a new error struct with the given kind, message, and options.

Parameters

kind - The kind of error, see kinds/0 for the available Kinds.

message - A string describing the error.

opts - A keyword list of additional options.

Options

:context - contextual information for better understanding the error. If provided, the context should be of type MscmpSystError.Types.Context.t/0.

:cause - The cause of the error, which can be an exception, an error tuple, or any other term.

Returns

Returns a struct of this error type with the error details.